% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_classes.R
\name{add_to_card.card}
\alias{add_to_card.card}
\title{#' @export
add_to_card.default <- function(x, trigger) {
c(x, trigger)
}}
\usage{
\method{add_to_card}{card}(x, trigger)
}
\description{
#' @export
add_to_card.default <- function(x, trigger) {
c(x, trigger)
}
}
