% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_classes.R
\name{add_to_card}
\alias{add_to_card}
\title{Jokers that trigger on certain cards are added as scoring to cards}
\usage{
add_to_card(x, trigger)
}
\arguments{
\item{x}{object to dispatch on}

\item{trigger}{a \code{chips}, \code{multp} or \code{multx} object}
}
\description{
Jokers that trigger on certain cards are added as scoring to cards
}
