% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balatro-package.R
\docType{package}
\name{balatro-package}
\alias{balatro-package}
\title{balatro: Package for calculating scores in the game Balatro}
\description{
Depending on your suit of jokers, level of hand and cards, calculate the score you will get
}
\author{
\strong{Maintainer}: Mathias Lerbech Jeppesen \email{mathiasljeppesen@outlook.com}

}
\keyword{internal}
