% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_classes.R
\name{chips}
\alias{chips}
\title{#' @export
chip_value.card <- function(x, ...) {
return(x$chip_value)
}}
\usage{
chips(
  x,
  card_trigger = NULL,
  foil = FALSE,
  holographic = FALSE,
  polychrome = FALSE
)
}
\description{
#' @export
chip_value.card_set <- function(x, ...) {
chips(sum(sapply(x, \(card) chip_value(card))))
}
}
