% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balatro.R
\name{balatro}
\alias{balatro}
\title{Title}
\usage{
balatro(
  base_score = balatro_score(),
  cards,
  jokers,
  debuff = NULL,
  ...,
  deck_format = build_deck()
)
}
\arguments{
\item{base_score}{a \code{balatro_score}}

\item{cards}{a \code{character} vector of card names}

\item{jokers}{a \code{vector} or \code{list} of function created with chips, multp
and multx functions}

\item{debuff}{a \code{character} specifying what type of cards are debuffed}

\item{...}{Nothing for now}

\item{deck_format}{a \code{character}}
}
\value{
a \code{numeric} with the score
}
\description{
Title
}
\examples{
balatro(cards = c("2h", "jc"), jokers = list(multp(10)), debuff = "even")
}
